occam-gateway/occam_gateway/requirements.txt
setuptools==69.2.0
django==4.2
djangorestframework==3.15.0
requests==2.29.0
drf-spectacular==0.24.1
coverage==7.4.3
django-oauth-toolkit==2.3.0
djangorestframework-api-key==3.*

# Evaluation
dinglehopper==0.9.6

# OCR
pdf2image==1.16.0

lxml~=5.3.0
pydantic~=1.10.18
symspellpy==6.7.8

celery[redis]==5.4.0
billiard==4.2.1
kombu==5.4.2
chardet==5.2.0